% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{mult  function to calculate the product of two arguments}
\usage{
mult(value1, value2)
}
\arguments{
\item{value1}{multiplicand 1}

\item{value2}{multiplicand 2}
}
\value{
numeric the product
}
\description{
mult  function to calculate the product of two arguments
}
\examples{
mult(3, 2)
# 6
}
